fix telegram schema
